translate utils
