+++ apps/meteor/client/providers/TranslationProvider.tsx
@@ -1,6 +1,14 @@
 import { useLocalStorage } from '@rocket.chat/fuselage-hooks';
 import languages from '@rocket.chat/i18n/dist/languages';
 import en from '@rocket.chat/i18n/src/locales/en.i18n.json';
+import {
+	addSprinfToI18n,
+	extractTranslationKeys,
+	applyCustomTranslations,
+	availableTranslationNamespaces,
+	defaultTranslationNamespace,
+	extractTranslationNamespaces,
+} from '@rocket.chat/i18n-utils';
 import { normalizeLanguage } from '@rocket.chat/tools';
 import type { TranslationContextValue } from '@rocket.chat/ui-contexts';
 import { useMethod, useSetting, TranslationContext } from '@rocket.chat/ui-contexts';
@@ -12,15 +20,7 @@ import { useEffect, useMemo } from 'react';
 import { I18nextProvider, initReactI18next, useTranslation } from 'react-i18next';
 
 import { getURL } from '../../app/utils/client';
-import {
-	i18n,
-	addSprinfToI18n,
-	extractTranslationKeys,
-	applyCustomTranslations,
-	availableTranslationNamespaces,
-	defaultTranslationNamespace,
-	extractTranslationNamespaces,
-} from '../../app/utils/lib/i18n';
+import { i18n } from '../../app/utils/lib/i18n';
 import { AppClientOrchestratorInstance } from '../apps/orchestrator';
 import { onLoggedIn } from '../lib/loggedIn';
 import { isRTLScriptLanguage } from '../lib/utils/isRTLScriptLanguage';
