+++ apps/meteor/server/lib/i18n.ts
@@ -1,6 +1,7 @@
 import i18nDict from '@rocket.chat/i18n';
+import { availableTranslationNamespaces, defaultTranslationNamespace, extractTranslationNamespaces } from '@rocket.chat/i18n-utils';
 
-import { availableTranslationNamespaces, defaultTranslationNamespace, extractTranslationNamespaces, i18n } from '../../app/utils/lib/i18n';
+import { i18n } from '../../app/utils/lib/i18n';
 
 void i18n.init({
 	lng: 'en',
