+++ packages/i18n-utils/.eslintrc.json
@@ -1,0 +1,4 @@
+{
+	"extends": ["@rocket.chat/eslint-config"],
+	"ignorePatterns": ["**/dist"]
+}
