+++ packages/i18n-utils/package.json
@@ -1,0 +1,29 @@
+{
+	"name": "@rocket.chat/i18n-utils",
+	"version": "0.0.2",
+	"private": true,
+	"devDependencies": {
+		"@types/chalk": "^2.2.4",
+		"@types/ejson": "^2.2.2",
+		"eslint": "~8.45.0",
+		"typescript": "~5.7.2"
+	},
+	"scripts": {
+		"lint": "eslint --ext .js,.jsx,.ts,.tsx .",
+		"lint:fix": "eslint --ext .js,.jsx,.ts,.tsx . --fix",
+		"build": "rm -rf dist && tsc -p tsconfig.json",
+		"dev": "tsc -p tsconfig.json --watch --preserveWatchOutput"
+	},
+	"main": "./dist/index.js",
+	"typings": "./dist/index.d.ts",
+	"files": [
+		"/dist"
+	],
+	"volta": {
+		"extends": "../../package.json"
+	},
+	"dependencies": {
+		"@rocket.chat/i18n": "workspace:^",
+		"i18next": "~23.4.9"
+	}
+}
