+++ packages/tools/src/index.ts
@@ -9,3 +9,4 @@ export * from './wrapExceptions';
 export * from './getLoginExpiration';
 export * from './converter';
 export * from './removeEmpty';
+export * from './isObject';
