+++ packages/tools/src/isObject.ts
@@ -1,0 +1,4 @@
+export function isObject(obj: unknown): obj is object {
+	const type = typeof obj;
+	return type === 'function' || (type === 'object' && !!obj);
+}
